Added logLaundryData route and timeoutCheck
the logLaundryData route takes a json sent by an esp and either adds or updates a machine document. This route also sends notifications when machines
turn from ON -> OFF.
timeoutCheck executes every 3 minutes and determines if 3 minutes have passed since a machine document has been updated. If it hasn't been updated,
an errorcode is added to the errorcode list of that machine